first part of 4.8
This exercise took me a long time to figure out. In the end it isn't really all that hard, but it took a while to view it the correct way.
I'm looking forward to writing out my solution to the second and third part, because I found them interesting, although they were pretty easy once I figured out 4.4 because I used basically the same ideas.